Deepen auth module: consolidate guards, responses, and session duration
- Add requireAuth/requireGuest guards used by _authed layout and login route
- Add loginWithPassword/logoutSession response builders used by API routes
- Unify session duration as a single SESSION_DURATION_SECONDS constant
- Route handlers now delegate auth logic instead of assembling internals

Closes #74
