+++ src/routes/_authed.tsx
@@ -1,12 +1,7 @@
-import { createFileRoute, Outlet, redirect } from '@tanstack/react-router'
-import { getIsAuthenticated } from '#/auth/server'
+import { createFileRoute, Outlet } from '@tanstack/react-router'
+import { requireAuth } from '#/auth/guards'
 
 export const Route = createFileRoute('/_authed')({
-  beforeLoad: async () => {
-    const isAuthenticated = await getIsAuthenticated()
-    if (!isAuthenticated) {
-      throw redirect({ to: '/login', search: { error: undefined } })
-    }
-  },
+  beforeLoad: requireAuth,
   component: () => <Outlet />,
 })
