+++ src/routes/api/auth/login.ts
@@ -1,9 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { env } from 'cloudflare:workers'
-import { verifyPassword, createSessionToken } from '#/auth/session'
-import { serializeSessionCookie } from '#/auth/cookies'
-
-const THIRTY_DAYS_SECONDS = 30 * 24 * 60 * 60
+import { loginWithPassword } from '#/auth/responses'
 
 export const Route = createFileRoute('/api/auth/login')({
   server: {
@@ -12,22 +8,14 @@ export const Route = createFileRoute('/api/auth/login')({
         const formData = await request.formData()
         const password = formData.get('password')
 
-        if (typeof password !== 'string' || !verifyPassword(password, env.APP_PASSWORD)) {
+        if (typeof password !== 'string') {
           return new Response(null, {
             status: 302,
             headers: { Location: '/login?error=invalid' },
           })
         }
 
-        const token = createSessionToken(env.APP_SECRET)
-
-        return new Response(null, {
-          status: 302,
-          headers: {
-            Location: '/',
-            'Set-Cookie': serializeSessionCookie(token, THIRTY_DAYS_SECONDS),
-          },
-        })
+        return loginWithPassword(password)
       },
     },
   },
